Add switch theme button in toolbar and rely it
