+++ controller/controller_session.py
@@ -10,6 +10,8 @@ import os
 import sys
 import subprocess
 
+import qdarkstyle
+
 import configs.hw_config as hw
 from controller.controller_main import MainController
 from ui.window_session import SessionWindow
@@ -42,6 +44,8 @@ class SessionController(SessionWindow):
         self.demo_gui_action.triggered.connect(self.runDemoGui)
         self.update_action.triggered.connect(self.update_hardware)
         self.close_action.triggered.connect(self.close)
+        self.switch_theme_action.triggered.connect(self.switch_theme)
+
 
         # Check if system is ready
         self.check_system()
@@ -242,6 +246,13 @@ class SessionController(SessionWindow):
         print('GUI closed successfully!')
         sys.exit()
 
+    def switch_theme(self):
+        self.is_dark_theme = not self.is_dark_theme
+        if self.is_dark_theme:
+            self.setStyleSheet(qdarkstyle.load_stylesheet_pyqt5())
+        else:
+            self.setStyleSheet("")  # Light theme: default Qt style
+
     def updateSessionDict(self):
         """
         Updates the session dictionary with the current session information.
@@ -250,21 +261,22 @@ class SessionController(SessionWindow):
         def get_text_or_placeholder(widget):
             return widget.text() if widget.text() else widget.placeholderText()
 
-        self.session = {
-            'project': self.tab_session.project_combo_box.currentText(),
-            'study': self.tab_session.study_combo_box.currentText(),
-            'side': self.tab_session.side_combo_box.currentText(),
-            'orientation': self.tab_session.orientation_combo_box.currentText(),
-            'subject_id': get_text_or_placeholder(self.tab_session.id_line_edit),
-            'study_id': get_text_or_placeholder(self.tab_session.idS_line_edit),
-            'subject_name': get_text_or_placeholder(self.tab_session.name_line_edit),
-            'subject_surname': get_text_or_placeholder(self.tab_session.surname_line_edit),
-            'subject_birthday': get_text_or_placeholder(self.tab_session.birthday_line_edit),
-            'subject_weight': get_text_or_placeholder(self.tab_session.weight_line_edit),
-            'subject_height': get_text_or_placeholder(self.tab_session.height_line_edit),
-            'scanner': get_text_or_placeholder(self.tab_session.scanner_line_edit),
-            'rf_coil': self.tab_session.rf_coil_combo_box.currentText(),
-            'software_version': get_text_or_placeholder(self.tab_session.software_line_edit),
-        }
+        self.session = {'project': self.tab_session.project_combo_box.currentText(),
+                        'study': self.tab_session.study_combo_box.currentText(),
+                        'side': self.tab_session.side_combo_box.currentText(),
+                        'orientation': self.tab_session.orientation_combo_box.currentText(),
+                        'subject_id': get_text_or_placeholder(self.tab_session.id_line_edit),
+                        'study_id': get_text_or_placeholder(self.tab_session.idS_line_edit),
+                        'subject_name': get_text_or_placeholder(self.tab_session.name_line_edit),
+                        'subject_surname': get_text_or_placeholder(self.tab_session.surname_line_edit),
+                        'subject_birthday': get_text_or_placeholder(self.tab_session.birthday_line_edit),
+                        'subject_weight': get_text_or_placeholder(self.tab_session.weight_line_edit),
+                        'subject_height': get_text_or_placeholder(self.tab_session.height_line_edit),
+                        'scanner': get_text_or_placeholder(self.tab_session.scanner_line_edit),
+                        'rf_coil': self.tab_session.rf_coil_combo_box.currentText(),
+                        'software_version': get_text_or_placeholder(self.tab_session.software_line_edit),
+                        'black_theme': self.is_dark_theme}
+
+        # Save session theme
 
         hw.b1Efficiency = hw.antenna_dict.get(self.session['rf_coil'], 1.0)
